[backport] Support fractional device pixel ratio in native Qt widget
When fractional DPI scaling is used, the VTK render window / interactor
has the wrong size because it's calculated from devicePixelRatio() which
is an integer API returning a rounded value.

Starting with Qt 5.6, there's a devicePixelRatioF() API that can return
a non-integer device pixel ratio. This makes sure the new API is used in
QVTKOpenGLWidget, when available.

The problem was first reported by me on the VTK developers mailing list:

    https://www.vtk.org/pipermail/vtk-developers/2018-February/035826.html