[backport] Improve vtkSelectVisiblePoints tolerance specification
Cherry-picked commit c5ef1ee4ee3 from main VTK repository.

Add the option to specify tolerance in world coordinates. It allows testing visibility when the tested points are visualized with glyphs. The glyphs always occlude the tested point, so a tolerance must be specified that is equal to the size of the point glyph. Size of the point glyph is known in world coordinate system, therefore it is necessary to specify the tolerance in world coordinate system.